requirements.txt
aliyun-python-sdk-cms==7.0.18
aliyun-python-sdk-core==2.13.30
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.3
django-mdeditor==0.1.18
django-ranged-response==0.2.0
django-redis==4.12.1
django-simple-captcha==0.5.13
idna==2.10
jmespath==0.10.0
Markdown==3.3.3
oauthlib==3.1.0
Pillow==8.0.1
pycparser==2.20
PyJWT==2.0.1
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2020.4
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
urllib3==1.26.2
